Making Int2ObjectMap configurable through a factory. This change allows to register a new Int2ObjectMap factory to the configuration while keeping the older implementation as the default. To set up a new factory clients of the library will have to create a factory and set it through the configuration. In addition clearing of the default configuration is now responsible for deciding how to clean itself while maintining implementation details encapsulated.
