+++ src/main/java/org/nustaq/serialization/util/DefaultFSTInt2ObjectMapFactory.java
@@ -1,0 +1,8 @@
+package org.nustaq.serialization.util;
+
+public class DefaultFSTInt2ObjectMapFactory implements FSTInt2ObjectMapFactory{
+    @Override
+    public <V> FSTInt2ObjectMap<V> createMap(int size) {
+        return new DefaultFSTInt2ObjectMap<>(size);
+    }
+}
