+++ src/main/java/org/nustaq/serialization/util/FSTInt2ObjectMapFactory.java
@@ -1,0 +1,5 @@
+package org.nustaq.serialization.util;
+
+public interface FSTInt2ObjectMapFactory {
+    <V> FSTInt2ObjectMap<V> createMap(int size);
+}
